requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
decorator==5.0.7
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Ext==0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==19.10.0
idna==3.1
infinity==1.5
intervals==0.9.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql==0.0.2
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.3
Pillow==8.2.0
pycparser==2.20
six==1.15.0
SQLAlchemy==1.4.6
validators==0.18.2
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-Components==0.10.5
